[refactor] 주문 코드 정리
