+++ src/main/java/kr/hhplus/be/server/domain/order/OrderItemEntity.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.domain.order;
 
 import jakarta.persistence.*;
-import kr.hhplus.be.server.domain.common.BaseTimeEntity;
+import kr.hhplus.be.server.domain.BaseTimeEntity;
 import kr.hhplus.be.server.domain.product.ProductEntity;
 import lombok.*;
